feat: use doubles for trigger values
Instead of longs

+++ src/main/java/com/brandwatch/kafka_pod_autoscaler/triggers/TriggerResult.java
@@ -2,8 +2,8 @@ package com.brandwatch.kafka_pod_autoscaler.triggers;
 
 import com.brandwatch.kafka_pod_autoscaler.v1alpha1.kafkapodautoscalerspec.Triggers;
 
-public record TriggerResult(Triggers trigger, long inputValue, long targetThreshold) {
+public record TriggerResult(Triggers trigger, double inputValue, double targetThreshold) {
     public int recommendedReplicas(int currentReplicaCount) {
-        return (int) Math.ceil(currentReplicaCount * (inputValue / (double) targetThreshold));
+        return (int) Math.ceil(currentReplicaCount * (inputValue / targetThreshold));
     }
 }
